requirements.txt
torch
torchvision
open-clip-torch
transformers
numpy
pandas
tqdm
h5py
Pillow
PyYAML
wandb
